requirements.txt
pydantic>=2.6.0,<3
PyYAML>=6.0.1,<7
pytest>=8.0.0,<9
